More logging refactoring
