+++ internal/run.go
@@ -1,8 +1,6 @@
 package internal
 
 import (
-	"context"
-
 	"os"
 	"os/exec"
 	"os/signal"
@@ -10,15 +8,14 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/ShowMax/go-fqdn"
-	"github.com/getsentry/sentry-go"
 	"github.com/go-kit/kit/log"
-	"github.com/kfdm/cron-me/internal/logger"
+
+	"github.com/kfdm/cron-me/internal/logging"
 )
 
 // signalWatcher handles our signals
 // https://github.com/Netflix/signal-wrapper/blob/master/main.go#L25
-func signalWatcher(ctx context.Context, cmd *exec.Cmd, logger log.Logger) {
+func signalWatcher(cmd *exec.Cmd, logger log.Logger) {
 	signalChan := make(chan os.Signal, 100)
 	signal.Notify(signalChan, syscall.SIGINT, syscall.SIGTERM)
 	signal := <-signalChan
@@ -53,17 +50,14 @@ func WrapReturn(cmd *exec.Cmd) int {
 
 // Run a command
 func Run(cmd *exec.Cmd) int {
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
 
-	logger := logger.NewFluentLogger()
+	logger := logging.NewFluentLogger()
 
 	user, _ := user.Current()
-	host := fqdn.Get()
 
-	logger.Log("tag", "cron.start", "User", user.Username, "Command", cmd.Args, "Host", host)
+	logger.Log("tag", "cron.start", "User", user.Username, "Command", cmd.Args)
 
-	go signalWatcher(ctx, cmd, logger)
+	go signalWatcher(cmd, logger)
 	cmd.Stdin = os.Stdin
 	cmd.Stdout = os.Stdout
 	cmd.Stderr = os.Stderr
@@ -73,20 +67,10 @@ func Run(cmd *exec.Cmd) int {
 	duration := time.Since(start)
 
 	if returncode == 0 {
-		logger.Log("tag", "cron.complete", "User", user.Username, "Command", cmd.Args, "Returncode", returncode, "Host", host, "Duration", duration.Seconds())
+		logger.Log("tag", "cron.complete", "User", user.Username, "Command", cmd.Args, "Returncode", returncode, "Duration", duration.Seconds())
 	} else {
-		logger.Log("tag", "cron.error", "User", user.Username, "Command", cmd.Args, "Returncode", returncode, "Host", host, "Duration", duration.Seconds())
-
-		err := sentry.Init(sentry.ClientOptions{})
-		if err == nil {
-			defer sentry.Flush(2 * time.Second)
-			sentry.ConfigureScope(func(scope *sentry.Scope) {
-				scope.SetLevel(sentry.LevelWarning)
-				scope.SetUser(sentry.User{Username: user.Username, ID: user.Uid})
-				scope.SetExtra("Returncode", returncode)
-			})
-			sentry.CaptureMessage(cmd.String())
-		}
+		logger.Log("tag", "cron.error", "User", user.Username, "Command", cmd.Args, "Returncode", returncode, "Duration", duration.Seconds())
+		logging.Sentry(user, cmd, returncode)
 	}
 
 	return returncode
